sequent_core

Module ballot

Ballot structures and helpers.

Structs

AreaAnnotations

Struct to hold annotations for an area.

AreaPresentation

Presentation settings for an area.

AuditableBallot

An auditable ballot.

AuditableBallotContest

An auditable contest on a ballot.

BallotStyle

Struct representing the ballot style, which includes information about the contests, areas, and presentation settings for a specific ballot.

Candidate

Candidate data structure.

CandidatePresentation

Presentation configuration for a candidate, including i18n, status, and display options.

CandidateUrl

URL and metadata for a candidate’s resource (e.g., website or image).

Contest

Contest data structure.

ContestPresentation

Presentation settings for a contest.

CustomUrls

Struct to hold custom URLs for the election event.

Election

Election data structure.

ElectionEventLanguageConf

Language configuration for an election event.

ElectionEventMaterials

Election event materials configuration. Materials configuration for an election event.

ElectionEventPresentation

Presentation configuration for an election event.

ElectionEventStatistics

Statistics related to an election event.

ElectionEventStatus

Status of the voting for the election event

ElectionPresentation

Presentation settings for an election event.

ElectionStatistics

Statistics related to an election, such as the number of emails and SMS sent.

ElectionStatus

Struct to hold the election status related to voting and tally.

HashableBallot

Hashable ballot.

HashableBallotContest

Contest data for hashable ballots.

PeriodDates

Struct to hold the first and last timestamps for each voting status change during a voting period.

PublicKeyConfig

Configuration for a public key.

RawHashableBallot

Raw hashable ballot data.

ReplicationChoice

Represents a choice in a contest.

ScheduledEventDates

Struct to hold the scheduled and stopped timestamps for scheduled events.

SignedContent

Content for a signed ballot, including the public key and signature.

SignedHashableBallot

Signed hashable ballot.

StringifiedPeriodDates

Struct to hold the stringified first and last timestamps for each voting status change during a voting period.

SubtypePresentation

Presentation settings for a candidate subtype.

TypePresentation

Presentation settings for a candidate type, including its subtypes.

VotingPeriodDates

Voting period dates.

VotingPortalCountdownPolicy

Configuration for the voting portal countdown Policy.

Weight

Struct to represent the weight of an area, which can be used in weighted voting systems.

Enums

AllowTallyStatus

Policy for allowing tally before voting period ends.

AuditButtonCfg

Configuration for the audit button in the voting portal.

CandidatesIconCheckboxPolicy

Policy for the icon used for candidate selection.

CandidatesOrder

Specifies the order in which candidates are displayed on the ballot.

CandidatesSelectionPolicy

Policy for candidate selection behavior

CastVoteGoldLevelPolicy

Policy for requiring gold level authentication when casting a vote.

ConsolidatedReportPolicy

Policy to determine if the consolidated report should be generated.

ContestEncryptionPolicy

Configuration for contest encryption policy.

ContestsOrder

Specifies the order in which contests are displayed on the ballot.

DecodedBallotsInclusionPolicy

Configuration for whether decoded ballots are included or not.

DelegatedVotingPolicy

Policy to determine if and when delegated voting is allowed.

EBlankVotePolicy

Policy for blank votes.

ECountdownPolicy

Policy for the voting portal countdown.

EDuplicatedRankPolicy

Policy for duplicated ranks in preferential voting.

EGracePeriodPolicy

Policy for grace period.

EInitializeReportPolicy

Policy for whether Initialize Report is required to start voting.

EOverVotePolicy

Policy for overvotes.

EPreferenceGapsPolicy

Policy for preference gaps in preferential voting.

ESecurityConfirmationPolicy

Policy for requiring security confirmation before voting.

EUnderVotePolicy

Policy for undervotes.

EarlyVotingPolicy

Policy for allowing or disallowing early voting.

ElectionsOrder

Policy for the order in which elections are displayed on the ballot.

Enrollment

configuration for whether enrollment is enabled or disabled.

InitReport

Policy for initialization report.

InvalidVotePolicy

Policy for handling invalid votes.

KeysCeremonyPolicy

Policy for whether executing key ceremonies at the election event level or election level.

LockedDown

Configuration for whether the election event is locked down or not.

ManualStartVotingPeriod

Policy for manually starting the voting period.

Otp

Configuration for whether OTP is enabled or disabled.

Publish

Configuration for whether able to publish.

ShowCastVoteLogs

Policy for showing or hiding the cast vote logs tab.

StartScreenTitlePolicy

Policy for the title shown on the start screen of the voting portal.

Tally

Policy for allowing tally before voting period ends.

TieBreakingPolicy

Policy to determine how ties are broken in the election.

VoterDigitalCertPolicy

Configuration for voter digital certificate policy.

VoterSigningPolicy

Configuration for voter signing policy.

VotingPeriodEnd

Policy for allowing voting period ends.

VotingStatus

Voting status.

VotingStatusChannel

Voting channels.

WeightedVotingPolicy

Policy to determine whether weighted voting is enabled.

Constants

TYPES_VERSION

Version number for ballot.

Functions

format_date

Helper method to format the date or return default.

format_date_opt

Helper method to format the date or return None.

get_ballot_bytes_for_signing

Get bytes of ballot content

sign_hashable_ballot_with_ephemeral_voter_signing_key

Errors

verify_ballot_signature

Errors

Type Aliases

Annotations

Annotations for ballots or contests, as key-value pairs.

I18nContent

Internationalized content map, keyed by language code.