sequent_core::ballot_codec::character_map

Trait CharacterMap

pub trait CharacterMap {
    // Required methods
    fn to_bytes(&self, s: &str) -> Result<Vec<u8>, String>;
    fn to_string(&self, bytes: &[u8]) -> Result<String, String>;
    fn base(&self) -> u64;
}

Trait for mapping between characters and bytes for ballot encoding.

Required Methods

fn to_bytes(&self, s: &str) -> Result<Vec<u8>, String>

Converts a string to a vector of bytes according to the map.

Errors

Returns an error if the string contains characters that cannot be mapped.

fn to_string(&self, bytes: &[u8]) -> Result<String, String>

Converts a byte slice to a string according to the map.

Errors

Returns an error if the byte slice contains bytes that cannot be mapped.

fn base(&self) -> u64

Returns the base (number of possible values) for this character map.