sequent_core::ballot_codec

Module checker

Structs

CheckerResult

Result of a ballot checker operation, containing errors and alerts.

Functions

check_blank_vote_policy

Checks the blank vote policy for a contest.

check_duplicated_rank_policy

Checks the duplicated rank policy for a contest.

check_invalid_vote_policy

Checks the invalid vote policy for a contest.

check_max_min_votes_policy

Checks the validity of max and min votes policy.

check_min_vote_policy

Checks if the number of selected candidates meets the minimum votes policy.

check_over_vote_policy

Checks the over-vote policy for a contest.

check_preference_gaps_policy

Checks the preference gaps policy for a contest.

check_under_vote_policy

Checks the under-vote policy for a contest.