sequent_core

Module interpret_plaintext

Plaintext ballot interpretation helpers.

Structs

ContestLayoutProperties

Struct representing the layout properties of a contest, including its state and whether it is sorted or ordered.

Enums

ContestState

Enum representing the different states of the contest UI.

Functions

check_is_blank

Checks if the given decoded contest is blank, meaning it has no explicit invalidity and all choices are unselected.

get_layout_properties

Determines the layout properties of a contest based on its counting algorithm.

get_points

@returns number of points this ballot is giving to this option