sequent_core::serialization

Module deserialize_with_path

Functions

deserialize_str

Deserialize a value of type T from a JSON string, tracking the path to any error.

deserialize_value

Deserialize a value of type T from a serde_json::Value, tracking the path to any error.