sequent_core::services

Module keycloak

Keycloak integration modules: admin client, permission, realm, role, and user management.

Structs

GroupInfo

KeycloakAdminClient

KeycloakAdminClient is a wrapper around the KeycloakAdmin client.

KeycloakLoginConfig

Configuration for logging into Keycloak using client credentials.

PubKeycloakAdmin

Struct to hold token and client.

PubKeycloakAdminToken

KeycloakAdminToken is a struct that represents the token response from Keycloak.

Enums

RoleAction

Constants

MULTIVALUE_USER_ATTRIBUTE_SEPARATOR

Functions

extract_realm_replacements

Extracts tenant_id and election_event_id replacements from a realm config.

get_auth_credentials

Acquire admin credentials from Keycloak using the provided login config.

get_client_credentials

Client Credentials OpenID Authentication flow. This enables servers to authenticate, without using a browser.

get_credentials_inner

Acquire credentials from Keycloak using the provided login config.

get_event_realm

Returns the event realm string for a tenant and election event.

get_tenant_realm

get_third_party_client_access_token

Authenticate a party client in keycloak with specific client credentials and tenant_id.

parse_realm

Parses a Keycloak realm string into tenant and optional event ID.

replace_realm_ids

Replaces UUIDs in a Keycloak realm JSON configuration while preserving specific IDs.