sequent_core::services::keycloak

Struct KeycloakAdminClient

pub struct KeycloakAdminClient {
    pub client: KeycloakAdmin,
}

KeycloakAdminClient is a wrapper around the KeycloakAdmin client.

Fields
client: KeycloakAdmin
Implementations

impl KeycloakAdminClient

pub async fn new() -> Result<KeycloakAdminClient>

Tries to read the token from the cache, if expired requests it to Keycloak.

Errors

Returns an error if credentials cannot be acquired or parsed.

pub async fn new_requested() -> Result<KeycloakAdminClient>

Creates a KeycloakAdminClient via fresh token requesting to Keycloak.

Errors

Returns an error if credentials cannot be acquired or parsed.

pub async fn pub_new() -> Result<PubKeycloakAdmin>

Creates a PubKeycloakAdmin with the admin token and client.

Errors

Returns an error if acquiring the Keycloak admin token fails.

impl KeycloakAdminClient

pub async fn list_permissions( self, realm: &str, search: Option<String>, limit: Option<usize>, offset: Option<usize>, ) -> Result<(Vec<Permission>, usize)>

List permissions for a realm, optionally filtered and paginated.

Errors

Returns an error if the Keycloak API call fails or if slicing fails due to invalid indices.

pub async fn set_role_permission( self, realm: &str, role_id: &str, permission_name: &str, ) -> Result<()>

Set a single role permission for a role in a realm.

Errors

Returns an error if the Keycloak API call fails.

pub async fn set_role_permissions( self, realm: &str, role_id: &str, permissions_name: &Vec<String>, ) -> Result<()>

Set multiple role permissions for a role in a realm.

Errors

Returns an error if the Keycloak API call fails.

pub async fn delete_role_permission( self, realm: &str, role_id: &str, permission_name: &str, ) -> Result<()>

Delete a single role permission from a role in a realm.

Errors

Returns an error if the Keycloak API call fails.

pub async fn delete_permission( self, realm: &str, permission_name: &str, ) -> Result<()>

Delete a permission from a realm.

Errors

Returns an error if the Keycloak API call fails.

pub async fn create_permission( self, realm: &str, permission: &Permission, ) -> Result<Permission>

Creates a permission in the given realm.

Errors

Returns an error if the Keycloak API call fails or the permission cannot be created.

impl KeycloakAdminClient

pub async fn get_realm( self, client: &PubKeycloakAdmin, board_name: &str, ) -> Result<RealmRepresentation, KeycloakError>

Gets the realm representation for a board.

Errors

Returns an error if the request fails or the token cannot be obtained.

pub async fn get_flow_executions( &self, client: &PubKeycloakAdmin, board_name: &str, execution_name: &str, ) -> Result<Vec<AuthenticationExecutionInfoRepresentation>, KeycloakError>

Gets the flow executions for a board and execution name.

Errors

Returns an error if the request fails or the token cannot be obtained.

pub async fn upsert_flow_execution( &self, client: &PubKeycloakAdmin, board_name: &str, execution_name: &str, json_execution_config: &str, ) -> Result<()>

Upserts a flow execution for a board.

Errors

Returns an error if the request fails or the token cannot be obtained.

pub async fn partial_import_realm_with_cleanup( &self, client: &PubKeycloakAdmin, tenant_id: &str, container_id: &str, realm_groups: Vec<GroupRepresentation>, realm_roles: Vec<RoleRepresentation>, if_resource_exists: &str, ) -> Result<()>

Partially imports a realm with cleanup.

Errors

Returns an error if the request fails or the token cannot be obtained.

pub async fn realm_delete( &self, client: &PubKeycloakAdmin, tenant_id: &str, delete_by: &str, id: &str, ) -> Result<(), KeycloakError>

Deletes a realm.

Errors

Returns an error if the request fails or the token cannot be obtained.

pub async fn create_new_group( &self, tenant_id: &str, group_name: &str, keycloak_client: &PubKeycloakAdmin, ) -> Result<Option<String>, KeycloakError>

Creates a new group in the realm.

Errors

Returns an error if the request fails or the token cannot be obtained.

pub async fn add_roles_to_group( &self, tenant_id: &str, keycloak_client: &PubKeycloakAdmin, group_id: &str, roles: &[RoleRepresentation], action: RoleAction, ) -> Result<(), KeycloakError>

Adds roles to a group in the realm.

Errors

Returns an error if the request fails or the token cannot be obtained.

pub async fn get_group_assigned_roles( &self, tenant_id: &str, group_id: &str, keycloak_client: &PubKeycloakAdmin, ) -> Result<Vec<RoleRepresentation>, Box<dyn Error>>

Gets the roles assigned to a group in Keycloak.

Errors

Returns an error if the request fails or the token cannot be obtained.

pub async fn update_group( &self, tenant_id: &str, group: &GroupRepresentation, ) -> Result<()>

Updates a group in Keycloak.

Errors

Returns an error if the request fails or the token cannot be obtained.

Panics

Panics if group.id is None.

pub async fn update_localization_texts_from_import( &self, imported_localization_texts: Option<HashMap<String, HashMap<String, String>>>, keycloak_client: &PubKeycloakAdmin, tenant_id: &str, ) -> Result<()>

Updates localization texts from imported data in Keycloak.

Errors

Returns an error if the request fails or the token cannot be obtained.

pub async fn upsert_realm( self, board_name: &str, json_realm_config: &str, tenant_id: &str, replace_ids: bool, display_name: Option<String>, election_event_id: Option<String>, ) -> Result<()>

Upserts a realm in Keycloak.

Errors

Returns an error if the request fails or the token cannot be obtained.

impl KeycloakAdminClient

pub async fn update_realm_attributes<S: BuildHasher + Send>( self, realm: &str, attributes: HashMap<String, String, S>, ) -> Result<()>

Update attributes into the realm’s existing Keycloak realm attributes.

Errors

Returns an error if the realm cannot be read or updated via the Keycloak Admin API.

impl KeycloakAdminClient

pub async fn list_roles( self, realm: &str, search: Option<String>, limit: Option<usize>, offset: Option<usize>, ) -> Result<(Vec<Role>, usize)>

Lists roles in the given realm.

Errors

Returns an error if the request fails or the API call is unsuccessful.

pub async fn list_user_roles( self: &KeycloakAdminClient, realm: &str, user_id: &str, ) -> Result<Vec<Role>>

Lists user roles in the given realm.

Errors

Returns an error if the request fails or the API call is unsuccessful.

pub async fn set_user_role( self: &KeycloakAdminClient, realm: &str, user_id: &str, role_id: &str, ) -> Result<()>

Sets a user role in the given realm.

Errors

Returns an error if the request fails or the API call is unsuccessful.

pub async fn delete_user_role( self, realm: &str, user_id: &str, role_id: &str, ) -> Result<()>

Deletes a user role in the given realm.

Errors

Returns an error if the request fails or the API call is unsuccessful.

pub async fn delete_role(self, realm: &str, role_id: &str) -> Result<()>

Deletes a role in the given realm.

Errors

Returns an error if the request fails or the API call is unsuccessful.

pub async fn create_role(self, realm: &str, role: &Role) -> Result<Role>

Creates a role in the given realm.

Errors

Returns an error if the request fails or the API call is unsuccessful.

pub async fn get_role_by_name(self, realm: &str, role: &Role) -> Result<Role>

Gets a role by name in the given realm.

Errors

Returns an error if the request fails or the API call is unsuccessful.

impl KeycloakAdminClient

pub async fn list_users( self, tenant_id: &str, election_event_id: &str, realm: &str, search: Option<String>, email: Option<String>, limit: Option<i32>, offset: Option<i32>, ) -> Result<(Vec<User>, i32)>

Lists users in the given realm.

Errors

Returns an error if the request fails or the API call is unsuccessful.

pub async fn get_user(&self, realm: &str, user_id: &str) -> Result<User>

Gets a user by id and given realm.

Errors

Returns an error if the request fails or the API call is unsuccessful.

pub async fn edit_user( self, realm: &str, user_id: &str, enabled: Option<bool>, attributes: Option<HashMap<String, Vec<String>>>, email: Option<String>, first_name: Option<String>, last_name: Option<String>, username: Option<String>, password: Option<String>, temporary: Option<bool>, ) -> Result<User>

Edits a user in the given realm.

Errors

Returns an error if the request fails or the API call is unsuccessful.

pub async fn edit_user_with_credentials( self, realm: &str, user_id: &str, enabled: Option<bool>, attributes: Option<HashMap<String, Vec<String>>>, email: Option<String>, first_name: Option<String>, last_name: Option<String>, username: Option<String>, credentials: Option<Vec<CredentialRepresentation>>, temporary: Option<bool>, ) -> Result<User>

Edits a user with credentials in the given realm.

Errors

Returns an error if the request fails or the API call is unsuccessful.

pub async fn delete_user(&self, realm: &str, user_id: &str) -> Result<()>

Deletes a user in the given realm.

Errors

Returns an error if the request fails or the API call is unsuccessful.

pub async fn create_user( self: &KeycloakAdminClient, realm: &str, user: &User, attributes: Option<HashMap<String, Vec<String>>>, groups: Option<Vec<String>>, ) -> Result<User>

Creates a user in the given realm.

Errors

Returns an error if the request fails or the API call is unsuccessful.

pub async fn get_user_profile_attributes( self: &KeycloakAdminClient, realm: &str, ) -> Result<Vec<UserProfileAttribute>>

Gets user profile attributes in the given realm.

Errors

Returns an error if the request fails or the API call is unsuccessful.

pub async fn get_user_groups( self: &KeycloakAdminClient, realm: &str, user_id: &str, ) -> Result<Vec<GroupInfo>>

Gets user groups for a user in the given realm.

Errors

Returns an error if the request fails or the API call is unsuccessful.

pub fn get_attribute_name(name: &Option<String>) -> Option<String>

pub fn get_formatted_attributes( attributes_res: &[UPAttribute], ) -> Vec<UserProfileAttribute>