sequent_core::services

Module reports

Structs

divide

eq

modulo

multiply

next

parse_i64

sum

Functions

expr_helper

Writes the first parameter as a string to the output.

format_date

Formats a date string using a dynamic format string.

format_datetime

Convert unix time to the given format.

format_dec_percentage

Formats a decimal as a percentage with two decimal places.

format_percentage

Formats a number as a percentage with two decimal places.

format_u64

Formats a u64 value with locale-specific separators.

helper_wrapper

helper_wrapper_or

inc

Increments an integer by 1 and writes it to output.

inc2

Increments an integer by 2 and writes it to output.

let_helper

Sets a variable in the Handlebars render context.

render_template

Renders a Handlebars template by name from a template map and variables map.

render_template_text

Renders a Handlebars template from a string and variables map.

sanitize_html

Sanitizes HTML input.

timestamp_to_rfc2822

Convert unix time to RFC2822 date and time format, like: Tue, 1 Jul 2003 10:52:37 +0200.

to_json

Serializes a parameter to JSON and writes it to output.