sequent_core

Module temp_path

Temporary file and path helpers for reports.

Constants

LOGO_TEMPLATE

The template html for the logo in the public assets.

PUBLIC_ASSETS_ELECTORAL_RESULTS_TEMPLATE_SYSTEM

The system template for the electoral results report.

PUBLIC_ASSETS_EML_BASE_TEMPLATE

The base template for EML documents.

PUBLIC_ASSETS_I18N_DEFAULTS

The default i18n JSON file for public assets.

PUBLIC_ASSETS_INITIALIZATION_TEMPLATE_SYSTEM

The system template for the initialization report.

PUBLIC_ASSETS_LOGO_IMG

The filename for the logo image in the public assets.

PUBLIC_ASSETS_QRCODE_LIB

The filename for the QR code library in the public assets.

PUBLIC_ASSETS_VELVET_BALLOT_IMAGES_TEMPLATE

The template for the ballot images page for the ballot images report.

PUBLIC_ASSETS_VELVET_BALLOT_IMAGES_TEMPLATE_SYSTEM

The system template for the ballot images page for the ballot images report.

PUBLIC_ASSETS_VELVET_MC_BALLOT_IMAGES_TEMPLATE

The template for the ballot images page for the MC ballot images report.

QR_CODE_TEMPLATE

The template for the QR code in the public assets.

VELVET_BALLOT_IMAGES_TEMPLATE_TITLE

The title for the ballot images template.

Functions

generate_temp_file

Generates a named temporary file with the given prefix and suffix.

get_file_size

Gets the file size for the given file path.

get_public_assets_path_env_var

Gets the public assets path from the environment variable.

read_temp_file

Reads the contents of a named temporary file.

read_temp_path

Reads the contents of a temporary file path.

write_into_named_temp_file

Errors