sequent_core::types

Module ceremonies

Ceremony-related types and enums.

Structs

KeysCeremonyStatus

Represents the state of a keys ceremony.

Log

Audit/event log entry for ceremonies, used for timestamped messages in ceremony workflows.

TallyCeremonyStatus

Represents the state of a tally ceremony

TallyElection

Represents the status of an election in a tally ceremony.

TallyResolution

Represents the resolution of a tally for a specific contest, including the selected candidate.

TallySessionDocuments

Holds paths to output documentss generated during a tally session

TallySessionResolution

Represents the resolution of a tally session, including details about the tie and how it was resolved.

TallySessionResolutionData

Data related to the resolution of a tie in a tally session.

TallyTrustee

Represents a participant in the tally ceremony.

Trustee

Represents a participant in the keys ceremony.

Enums

CeremoniesPolicy

Configures whether ceremonies are conducted manually or automatically.

CountingAlgType

Specifies the counting algorithm used for tallying votes.

KeysCeremonyExecutionStatus

Represents the status of the keys ceremony execution.

ScopeOperation

Specifies whether a tally operation is scoped to an area or a contest.

TallyElectionStatus

Status of the tally execution for an election (waiting, mixing, decrypting, success, error).

TallyExecutionStatus

Represents the status of the tally execution.

TallyOperation

Specifies the operation performed during tallying.

TallySessionResolutionStatus

Status of a tally session resolution.

TallySessionResolutionType

Types of resolutions that can be applied to a tally session.

TallyTrusteeStatus

Status of a trustee in the tally ceremony

TallyType

Distinguishes the type of tally report generated

TieBreakingMethod

Specifies the counting algorithm used for tallying votes in an election. Specifies the method used to break ties in preferential voting.

TrusteeStatus

Represents the status of a trustee in the keys ceremony.