sequent_core::types

Module keycloak

Keycloak integration types.

Structs

Permission

Represents a permission in Keycloak.

Role

Represents a role in Keycloak.

UPAttributePermissions

Permissions for editing and viewing user profile attributes.

UPAttributeRequired

Required roles and scopes for a user profile attribute.

UPAttributeSelector

Selector for user profile attribute scopes.

User

Represents a user in Keycloak with profile and voting information.

UserArea

Represents an area assigned to a user.

UserProfileAttribute

Represents a user profile attribute in Keycloak.

VotesInfo

Information about votes cast by a user in an election.

Constants

AREA_ID_ATTR_NAME

Attribute name for area ID.

ATTR_RESET_VALUE

Value used to reset an attribute.

AUTHORIZED_ELECTION_IDS_NAME

Attribute name for authorized election IDs.

CERTIFICATES_IDP_ALIAS

Identity-provider alias for the digital-certificates IdP configured on tenant realms.

DATE_OF_BIRTH

Attribute name for date of birth.

DISABLE_COMMENT

A voter can be disabled:

DISABLE_REASON_DELETE_CALL

Reason string for disabling a voter via Datafix delete-voter endpoint.

DISABLE_REASON_MARKVOTED_CALL

Reason string for disabling a voter via Datafix mark-voted call.

FIRST_NAME

Attribute name for first name.

LAST_NAME

Attribute name for last name.

MOBILE_PHONE_ATTR_NAME

Attribute name for mobile phone number.

PERMISSION_LABELS

Attribute name for permission labels.

PERMISSION_TO_EDIT

Permission name for editing.

REALM_ATTR_VOTER_CERTIFICATE_POLICY

Realm attribute key storing the voter certificate policy.

TENANT_ID_ATTR_NAME

Attribute name for tenant ID.

VOTED_CHANNEL

If there is a call to Datafix mark-voted, we disable the voter and set this value to signal the channel e.g. PHONE, POST, etc.

VOTED_CHANNEL_INTERNET_VALUE

Value for internet voting channel.