sequent_core::util

Module date_time

Constants

PHILIPPINO_TIMEZONE

Functions

generate_timestamp

Generate a timestamp string based on the provided time zone, date format, and date time.

get_date_and_time

Get the current date and time in RFC 3339 format.

get_system_timezone

Get the current system time zone.

verify_date_format_ymd

Check if the date is correct, format must be YYYY-MM-DD. Date in the future is not valid.