sequent_core::util

Module temp_path

Functions

generate_temp_file

Generates a named temporary file with the given prefix and suffix.

get_file_size

Gets the file size for the given file path.

get_public_assets_path_env_var

Gets the public assets path from the environment variable.

read_temp_file

Reads the contents of a named temporary file.

read_temp_path

Reads the contents of a temporary file path.

write_into_named_temp_file

Writes data into a named temp file. The temp file will have the specificed prefix and suffix.