sequent_core

Module wasm

WebAssembly bindings exported to frontend packages. This module contains the WebAssembly bindings for the Sequent Core library. It includes TypeScript interfaces and enums that are used to interact with the library from JavaScript/TypeScript code.

Modules

areas

WebAssembly bindings exported to frontend packages for accessing area information.

templates

wasm_bindgen for template related types and interfaces.

wasm

WebAssembly bindings exported to frontend packages.

wasm_hasura_types

wasm_bindgen for hasura related types and interfaces.

wasm_interpret_plaintext

wasm_bindgen for interpret plaintext related types and interfaces.

wasm_keycloak

wasm_bindgen for Keycloak related types and interfaces.

wasm_permissions

wasm_bindgen for permission related types and interfaces.

wasm_plaintext

wasm_bindgen for plaintext related types and interfaces.