sequent_core::wasm::wasm

Function sign_hashable_ballot_with_ephemeral_voter_signing_key_js

pub fn sign_hashable_ballot_with_ephemeral_voter_signing_key_js(
    ballot_id: JsValue,
    election_id: JsValue,
    content: JsValue,
) -> Result<JsValue, JsValue>

Signs a hashable ballot with an ephemeral voter signing key.

Errors

Returns an error if parsing, conversion, or signing fails.